+++ admin/includes/view_all_posts.php
@@ -1,10 +1,12 @@
 <?php 
 
+    include("delete_modal.php");
+
     //this post global is for making sure, that the bulk options that user pick for the post status could be updated here  
     if(isset($_POST['checkBoxArray']))
     {
         foreach ($_POST['checkBoxArray'] as $postValueId) {
-            $bulk_options = $_POST['bulk_options'];
+            $bulk_options = escape($_POST['bulk_options']);
 
             switch ($bulk_options) {
                 case 'Published':
@@ -34,19 +36,19 @@
                     $select_post_query = mysqli_query($connection, $query);
 
                     while ($row = mysqli_fetch_array($select_post_query)) {
-                        $post_author = $row['post_author'];
-                        $post_title = $row['post_title'];
-                        $post_category_id = $row['post_category_id'];
-                        $post_status = $row['post_status'];
+                        $post_user = escape($row['post_user']);
+                        $post_title = escape($row['post_title']);
+                        $post_category_id = escape($row['post_category_id']);
+                        $post_status = escape($row['post_status']);
                         $post_image = $row['post_image'];
-                        $post_tag = $row['post_tag'];
-                        $post_date = $row['post_date'];
-                        $post_content = $row['post_content'];
+                        $post_tag = escape($row['post_tag']);
+                        $post_date = escape($row['post_date']);
+                        $post_content = escape($row['post_content']);
                     }
 
-                    $query = "INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_image, post_content, post_tag,  post_status) ";
+                    $query = "INSERT INTO posts(post_category_id, post_title, post_user, post_date, post_image, post_content, post_tag,  post_status) ";
 
-                    $query .= "VALUES({$post_category_id},'$post_title','$post_author', now(),'{$post_image}','{$post_content}','{$post_tag}','{$post_status}')";
+                    $query .= "VALUES({$post_category_id},'$post_title','$post_user', now(),'{$post_image}','{$post_content}','{$post_tag}','{$post_status}')";
 
                     $copy_query = mysqli_query($connection, $query);
                     confirmQuery($copy_query);
@@ -82,7 +84,7 @@
 
     <div id="bulkOptionsButton" class="col-xs-4">
         <input type="submit" name="submit" class="btn btn-success " value="Apply" title="Apply Bulk Action">
-        <a class="btn btn-primary" href="posts.php?source=add_post" title="Add New Post">Add New</a>
+        <a class="btn btn-primary" href="posts.php?source=add_post" title="Add New Post"><i class="fa fa-plus"></i> Add New</a>
     </div> <br>
 
 
@@ -90,7 +92,7 @@
         <tr>
             <th><input  type="checkbox" id="selectAllBoxes" name="selectAllBoxes"></th>
             <th>Id</th>
-            <th>Author</th>
+            <th>Author (Users)</th>
             <th>Title</th>
             <th>Category</th>
             <th>Status</th>
@@ -114,23 +116,28 @@
 
     while ($row = mysqli_fetch_assoc($select_posts)) 
     { //amek and tukarkan column kepada key, and anak2 column as value dia s
-        $post_id = $row['post_id'];
-        $post_author = $row['post_author'];
-        $post_title = $row['post_title'];
-        $post_category_id = $row['post_category_id'];
-        $post_status = $row['post_status'];
-        $post_image = $row['post_image'];
-        $post_tag = $row['post_tag'];
-        $post_comment_count = $row['post_comment_count'];
-        $post_date = $row['post_date'];
-        $post_view_count = $row['post_view_count'];
+        $post_id = escape($row['post_id']);
+        $post_user = escape($row['post_user']);
+        $post_title = escape($row['post_title']);
+        $post_category_id = escape($row['post_category_id']);
+        $post_status = escape($row['post_status']);
+        $post_image = escape($row['post_image']);
+        $post_tag = escape($row['post_tag']);
+        $post_comment_count = escape($row['post_comment_count']);
+        $post_date = escape($row['post_date']);
+        $post_view_count = escape($row['post_view_count']);
 
         echo "<tr>";
         ?>
             <td><input class="checkBoxes" type="checkbox"  name="checkBoxArray[]" value="<?php echo $post_id ?>"></td>
         <?php
         echo "<td> $post_id </td>";
-        echo "<td> $post_author </td>";
+
+
+        echo "<td> $post_user </td>";
+        
+
+
         echo "<td> $post_title </td>";
 
 
@@ -140,8 +147,8 @@
          $select_categories_id = mysqli_query($connection, $query); 
 
          while ($row = mysqli_fetch_assoc( $select_categories_id )) { 
-         $cat_id = $row['cat_id'];
-         $cat_title = $row['cat_title'];
+         $cat_id = escape($row['cat_id']);
+         $cat_title = escape($row['cat_title']);
 
 
 
@@ -182,9 +189,9 @@
                   <ul class='dropdown-menu'>
                     <li><a href='../post.php?p_id={$post_id}' title='View Post'> <i class='fa fa-eye'></i> View</a></li>
                     <li class='divider'></li>
-                    <li><a href='posts.php?source=edit_post&p_id={$post_id}' title='Edit Post'><i class='fa fa-pencil'></i> Edit</a></li>
+                    <li><a  href='posts.php?source=edit_post&p_id={$post_id}' title='Edit Post'><i class='fa fa-pencil'></i> Edit</a></li>
                     <li class='divider'></li>
-                    <li><a onClick=\"javascript: return confirm('Are you sure you want to delete?');\" href='posts.php?delete={$post_id}' title='Delete Post'><i class='fa fa-trash'></i> Delete</a></li>
+                    <li><a rel='$post_id' href='javascript:void(0)' class='delete_link' title='Delete Post'><i class='fa fa-trash'></i> Delete</a></li>
                   </ul>
                 </div> 
             </td>";
@@ -204,39 +211,69 @@
 <?php 
 
 if (isset($_GET['delete'])) {
-        $the_post_id = $_GET['delete'];
 
-
-        $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
-        $deleteQuery = mysqli_query($connection, $query);
-
-        if(!$deleteQuery)
+    if(isset($_SESSION['user_role']))
+    {
+        if($_SESSION['user_role'] == 'Admin')
         {
-            die('QUERY FAILED' . mysqli_error($connection));
+            $the_post_id =  escape( $_GET['delete']);
+
+
+            $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
+            $deleteQuery = mysqli_query($connection, $query);
+
+            if(!$deleteQuery)
+            {
+                die('QUERY FAILED' . mysqli_error($connection));
+            }
+
+
+             header("Location: posts.php");       
         }
-
-
-         header("Location: posts.php");
     }
+
+}
 
 if (isset($_GET['reset'])) {
-        $the_post_id = $_GET['reset'];
 
-
-        $query = "UPDATE posts SET post_view_count = 0 WHERE post_id =" . mysqli_real_escape_string($connection, $the_post_id) . "";
-        $resetQuery = mysqli_query($connection, $query);
-
-        if(!$resetQuery)
+    if(isset($_SESSION['user_role']))
+    {
+        if($_SESSION['user_role'] == 'Admin')
         {
-            die('QUERY FAILED' . mysqli_error($connection));
+
+            $the_post_id =  escape( $_GET['reset']);
+
+
+            $query = "UPDATE posts SET post_view_count = 0 WHERE post_id =" . escape($the_post_id) . "";
+            $resetQuery = mysqli_query($connection, $query);
+
+            if(!$resetQuery)
+            {
+                die('QUERY FAILED' . mysqli_error($connection));
+            }
+
+
+             header("Location: posts.php");
         }
-
-
-         header("Location: posts.php");
     }
 
-
-
-
-
+}
  ?>
+
+<script>
+
+    $(document).ready(function(){
+
+        $(".delete_link").on('click', function(){
+
+            var id = $(this).attr("rel");
+            var delete_url = "posts.php?delete=" + id + " ";
+
+
+            $(".modal_delete_link").attr("href", delete_url); 
+
+            $("#delModal").modal("show")
+        });
+    });
+
+</script>
